pkg/vminfo: extract feature parsing logic from pkg/csource
Extract the command-line feature flag parsing logic (used for -enable
and -disable flags) out of pkg/csource and into a new subpackage
pkg/vminfo/features.

This breaks the dependency of tools/syz-execprog on pkg/csource, keeping
the host execution environment configuration cleanly separated from the
C-source generation codebase.
